+++ dotnet/Data/Constants.cs
@@ -33,6 +33,7 @@ namespace AvailabilityNotify.Data
 
         public const string BUCKET = "availability-notify";
         public const string LOCK = "availability-notify-lock";
+        public const string UNSENT_CHECK = "check-unsent";
 
         public const string DATA_ENTITY = "notify";
         public const string SCHEMA = "notify";
@@ -52,6 +53,42 @@ namespace AvailabilityNotify.Data
         {
             public const string CannotBeDelivered = "cannotBeDelivered";
             public const string Available = "available";
+        }
+
+        public static class Domain
+        {
+            public const string Fulfillment = "Fulfillment";
+            public const string Marketplace = "Marketplace";
+        }
+
+        public static class VtexOrderStatus
+        {
+            public const string OrderCreated = "order-created";
+            public const string OrderCompleted = "order-completed";
+            public const string OnOrderCompleted = "on-order-completed";
+            public const string PaymentPending = "payment-pending";
+            public const string WaitingForOrderAuthorization = "waiting-for-order-authorization";
+            public const string ApprovePayment = "approve-payment";
+            public const string PaymentApproved = "payment-approved";
+            public const string PaymentDenied = "payment-denied";
+            public const string RequestCancel = "request-cancel";
+            public const string WaitingForSellerDecision = "waiting-for-seller-decision";
+            public const string AuthorizeFullfilment = "authorize-fulfillment";
+            public const string OrderCreateError = "order-create-error";
+            public const string OrderCreationError = "order-creation-error";
+            public const string WindowToCancel = "window-to-cancel";
+            public const string ReadyForHandling = "ready-for-handling";
+            public const string StartHanding = "start-handling";
+            public const string Handling = "handling";
+            public const string InvoiceAfterCancellationDeny = "invoice-after-cancellation-deny";
+            public const string OrderAccepted = "order-accepted";
+            public const string Invoice = "invoice";
+            public const string Invoiced = "invoiced";
+            public const string Replaced = "replaced";
+            public const string CancellationRequested = "cancellation-requested";
+            public const string Cancel = "cancel";
+            public const string Canceled = "canceled";
+            public const string Cancelled = "cancelled";
         }
     }
 }