+++ dotnet/Services/IVtexAPIService.cs
@@ -9,10 +9,12 @@ namespace AvailabilityNotify.Services
         Task<bool> AvailabilitySubscribe(string email, string sku, string name, string locale, SellerObj sellerObj);
         Task<bool> ProcessNotification(AffiliateNotification notification);
         Task<bool> ProcessNotification(BroadcastNotification notification);
+        Task ProcessNotification(AllStatesNotification notification);
         Task<bool> VerifySchema();
         Task<bool> CreateDefaultTemplate();
         Task<List<string>> ProcessAllRequests();
         Task<List<string>> ProcessUnsentRequests();
+        Task CheckUnsentNotifications();
         Task<CartSimulationResponse> CartSimulation(CartSimulationRequest cartSimulationRequest, RequestContext requestContext);
     }
 }