+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Periodically process all unsent notifications
+
 ## [1.6.7] - 2022-05-06
 
 ### Changed